Allow stream to reconnect after disconnection
After a disconnection (not triggered by user) socketIO library was only
connected to stream server but not subscribing to streams that we were
subscribed before. The 3rd party library allows us to use namespace and
specify also what to do in case of different events. So with this commit
we add the feature to keep track of stream subscriptions and resuscribe
after reconnection. That would also tackle issues like #43